Moved multiline reader to TextLoaderCursor, and added return when multiline doesn't close quoting
